Moved API test
