+++ bee-ternary/src/lib.rs
@@ -660,18 +660,16 @@ impl<T: RawEncoding + ?Sized> ToOwned for Trits<T> {
 }
 
 impl<T: RawEncoding + ?Sized> fmt::Display for Trits<T> {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		write!(f, "[")?;
-		for (i, t) in self.iter().enumerate() {
-			if i != 0 {
-				write!(f, ", ")?;
-			}
-			write!(f, "{}", t)?;
-
-		}
-		write!(f, "]")
-
-	}
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "[")?;
+        for (i, t) in self.iter().enumerate() {
+            if i != 0 {
+                write!(f, ", ")?;
+            }
+            write!(f, "{}", t)?;
+        }
+        write!(f, "]")
+    }
 }
 
 /// A buffer containing trits.
@@ -821,9 +819,9 @@ impl<T: RawEncodingBuf> fmt::Debug for TritBuf<T> {
 }
 
 impl<T: RawEncodingBuf> fmt::Display for TritBuf<T> {
-	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		write!(f, "{}", self.as_slice())
-	}
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.as_slice())
+    }
 }
 
 impl<T: RawEncodingBuf> Borrow<Trits<T::Slice>> for TritBuf<T> {
